Changed the code for generating and displaying final report to clear resources like streams and files to prevent memory leaks. Fixed a bug with bringing in the external node added to a quick-entered sample on merging the sample on a login screen.

+++ src/org/openelis/gwt/server/ReportServlet.java
@@ -36,6 +36,7 @@ public class ReportServlet extends HttpServlet {
 
         key = null;
         out = null;
+        path = null;
         try {
             /*
              * make sure the requested file is in user session (we have
@@ -53,7 +54,6 @@ public class ReportServlet extends HttpServlet {
                 return;
             }
 
-            path = null;
             try {
                 path = Paths.get(status.getPath());
             } catch (InvalidPathException e) {
@@ -83,15 +83,20 @@ public class ReportServlet extends HttpServlet {
 
             out = resp.getOutputStream();
             Files.copy(path, out);
-            Files.delete(path);
         } catch (Exception e) {
             log.log(Level.SEVERE, e.getMessage(), e);
             throw new ServletException(e.getMessage());
         } finally {
-            if (out != null)
-                out.close();
-            if (key != null)
-                req.getSession().removeAttribute(key);
+            try {
+                if (out != null)
+                    out.close();
+                if (key != null)
+                    req.getSession().removeAttribute(key);
+                if (path != null)
+                    Files.delete(path);
+            } catch (Exception e) {
+                log.log(Level.SEVERE, e.getMessage(), e);
+            }
         }
     }
